Restore jit(_kernel)_base hierarchy
