+++ src/plugins/intel_cpu/src/nodes/kernels/grid_sample.hpp
@@ -4,7 +4,7 @@
 
 #pragma once
 
-#include "jit_kernel_base.hpp"
+#include "jit_kernel.hpp"
 #include "ie_precision.hpp"
 #include <set>
 
@@ -59,19 +59,11 @@ enum coord {
     w, h
 };
 
-class GridSampleKernelBase: public JitKernelBase {
+class GridSampleKernelBase : public jit_kernel<GridSampleKernelConfParams, GridSamplesKernelExecArgs> {
 public:
-    void (*ker_)(const GridSamplesKernelExecArgs *);
-    void operator()(const GridSamplesKernelExecArgs *args) {
-        assert(ker_);
-        ker_(args);
-    }
     explicit GridSampleKernelBase(const char* name, x64::cpu_isa_t max_cpu_isa, const GridSampleKernelConfParams& jcp)
-        : JitKernelBase(name, max_cpu_isa),
-          ker_(nullptr),
-          jcp(jcp) {}
+        : jit_kernel(name, max_cpu_isa, jcp) {}
 
-    virtual void create_ker() = 0;
     uint64_t getVecLen() {
         return vlen;
     }
@@ -83,7 +75,6 @@ public:
     }
 
 protected:
-    GridSampleKernelConfParams jcp;
     uint64_t vlen         = 16lu;
     uint64_t dataTypeSize = 1lu;
     uint64_t gridTypeSize = 1lu;
@@ -98,7 +89,6 @@ public:
 
     explicit GridSampleKernel(const GridSampleKernelConfParams& jcp);
 
-    void create_ker() override;
     void generate_impl() override;
 
     using Vmm   = typename dnnl::impl::utils::conditional3<isa == dnnl::impl::cpu::x64::avx512_core, Xbyak::Zmm,
