+++ src/plugins/intel_cpu/src/nodes/kernels/jit_kernel_base.cpp
@@ -4,13 +4,11 @@
 
 #include "jit_kernel_base.hpp"
 
-using namespace ov;
-using namespace intel_cpu;
-using namespace dnnl::impl::cpu;
+using namespace ov::intel_cpu;
 using namespace dnnl::impl::cpu::x64;
 
 
-void JitKernelBase::generate() {
+void jit_kernel_base::generate() {
     this->preamble();
 
     createRegistersPool();
@@ -28,164 +26,7 @@ void JitKernelBase::generate() {
     }
 }
 
-void JitKernelBase::uni_vfmsub132ps(const Xbyak::Xmm& vDst,
-                                    const Xbyak::Xmm& vSrc,
-                                    const Xbyak::Operand& op) {
-    if (isValidIsa(x64::avx2)) {
-        vfmsub132ps(vDst, vSrc, op);
-    } else if (isValidIsa(x64::avx)) {
-        assert(vDst.getIdx() != vSrc.getIdx());
-        vmulps(vDst, vDst, op);
-        vsubps(vDst, vDst, vSrc);
-    } else {
-        assert(vDst.getIdx() != vSrc.getIdx());
-        mulps(vDst, op);
-        subps(vDst, vSrc);
-    }
-}
-
-void JitKernelBase::uni_vfnmadd132ps(const Xbyak::Xmm& vDst,
-                                     const Xbyak::Xmm& vSrc,
-                                     const Xbyak::Operand& op) {
-    if (isValidIsa(x64::avx2)) {
-        vfnmadd132ps(vDst, vSrc, op);
-    } else if (isValidIsa(x64::avx)) {
-        assert(vDst.getIdx() != vSrc.getIdx());
-        vmulps(vDst, vDst, op);
-        vsubps(vDst, vSrc, vDst);
-    } else {
-        assert(vDst.getIdx() != vSrc.getIdx());
-        mulps(vDst, op);
-        subps(vSrc, vDst);
-        movups(vDst, vSrc);
-    }
-}
-
-void JitKernelBase::uni_vfmsub231ps(const Xbyak::Xmm& vDst,
-                                    const Xbyak::Xmm& vSrc,
-                                    const Xbyak::Operand& op) {
-    if (isValidIsa(x64::avx2)) {
-        vfmsub231ps(vDst, vSrc, op);
-    } else if (isValidIsa(x64::avx)) {
-        assert(!vDst.isEqualIfNotInherited(op));
-        vmulps(vSrc, vSrc, op);
-        vsubps(vDst, vSrc, vDst);
-    } else {
-        assert(!vDst.isEqualIfNotInherited(op));
-        mulps(vSrc, op);
-        subps(vSrc, vDst);
-        movups(vDst, vSrc);
-    }
-}
-
-void JitKernelBase::uni_vpaddd(const Xbyak::Ymm& vDst,
-                               const Xbyak::Ymm& vSrc,
-                               const Xbyak::Operand& op) {
-    if (isValidIsa(x64::avx2)) {
-        vpaddd(vDst, vSrc, op);
-    } else if (isValidIsa(x64::avx)) {
-        Xbyak::Xmm xmmDst(vDst.getIdx());
-        vmovups(vDst, vSrc);
-        if (op.isYMM()) {
-            Xbyak::Ymm ymmOp(op.getIdx());
-            Xbyak::Xmm xmmOp(op.getIdx());
-            paddd(xmmDst, xmmOp);
-            vperm2f128(vDst, vDst, vDst, 0x1);
-            vperm2f128(ymmOp, ymmOp, ymmOp, 0x1);
-            paddd(xmmDst, xmmOp);
-            vperm2f128(vDst, vDst, vDst, 0x1);
-            vperm2f128(ymmOp, ymmOp, ymmOp, 0x1);
-        } else if (op.isMEM()) {
-            const int vlen = x64::cpu_isa_traits<x64::sse41>::vlen;
-            paddd(xmmDst, op.getAddress());
-            vperm2f128(vDst, vDst, vDst, 0x1);
-            paddd(xmmDst, ptr[op.getAddress().getRegExp() + vlen]);
-            vperm2f128(vDst, vDst, vDst, 0x1);
-        } else {
-            IE_THROW() << "Not supported operand type.";
-        }
-    } else if (isValidIsa(x64::sse41)) {
-        assert(vDst.getIdx() != vSrc.getIdx());
-        paddd(vDst, op);
-    } else {
-        IE_THROW() << "Not defined behavior for instruction 'vpaddd' in current instructions set.";
-    }
-}
-
-void JitKernelBase::uni_vpsubd(const Xbyak::Ymm& vDst,
-                               const Xbyak::Ymm& vSrc,
-                               const Xbyak::Operand& op) {
-    if (isValidIsa(x64::avx2)) {
-        vpsubd(vDst, vSrc, op);
-    } else if (isValidIsa(x64::avx)) {
-        Xbyak::Xmm xmmDst(vDst.getIdx());
-        vmovups(vDst, vSrc);
-        if (op.isYMM()) {
-            Xbyak::Ymm ymmOp(op.getIdx());
-            Xbyak::Xmm xmmOp(op.getIdx());
-            psubd(xmmDst, xmmOp);
-            vperm2f128(vDst, vDst, vDst, 0x1);
-            vperm2f128(ymmOp, ymmOp, ymmOp, 0x1);
-            psubd(xmmDst, xmmOp);
-            vperm2f128(vDst, vDst, vDst, 0x1);
-            vperm2f128(ymmOp, ymmOp, ymmOp, 0x1);
-        } else if (op.isMEM()) {
-            const int vlen = x64::cpu_isa_traits<x64::sse41>::vlen;
-            psubd(xmmDst, op.getAddress());
-            vperm2f128(vDst, vDst, vDst, 0x1);
-            psubd(xmmDst, ptr[op.getAddress().getRegExp() + vlen]);
-            vperm2f128(vDst, vDst, vDst, 0x1);
-        } else {
-            IE_THROW() << "Not supported operand type.";
-        }
-    } else if (isValidIsa(x64::sse41)) {
-        assert(vDst.getIdx() != vSrc.getIdx());
-        psubd(vDst, op);
-    } else {
-        IE_THROW() << "Not defined behavior for instruction 'vpsubd' in current instructions set.";
-    }
-}
-
-void JitKernelBase::uni_vdivps(const Xbyak::Xmm& vDst,
-                               const Xbyak::Operand& op1,
-                               const Xbyak::Operand& op2) {
-    if (isValidIsa(x64::avx)) {
-        vdivps(vDst, op1, op2);
-    } else {
-        if (!vDst.isEqualIfNotInherited(op1)) {
-            movups(vDst, op1);
-        }
-        divps(vDst, op2);
-    }
-}
-
-void JitKernelBase::uni_vandps(const Xbyak::Xmm& vDst,
-                               const Xbyak::Xmm& vSrs,
-                               const Xbyak::Operand &op) {
-    if (isValidIsa(x64::avx)) {
-        vandps(vDst, vSrs, op);
-    } else {
-        if (!vDst.isEqualIfNotInherited(vSrs)) {
-            movups(vDst, vSrs);
-        }
-        andps(vDst, op);
-    }
-}
-
-void JitKernelBase::uni_vandnps(const Xbyak::Xmm& vDst,
-                                const Xbyak::Xmm& vSrs,
-                                const Xbyak::Operand &op) {
-    if (isValidIsa(x64::avx)) {
-        vandnps(vDst, vSrs, op);
-    } else {
-        if (!vDst.isEqualIfNotInherited(vSrs)) {
-            movups(vDst, vSrs);
-        }
-        andnps(vDst, op);
-    }
-}
-
-void JitKernelBase::emu_vgatherdps(const Xbyak::Xmm& xmm_val,
+void jit_kernel_base::emu_vgatherdps(const Xbyak::Xmm& xmm_val,
                                    const Xbyak::Reg64& reg_addr,
                                    const Xbyak::Xmm& xmm_index,
                                    const int& scale,
@@ -195,7 +36,7 @@ void JitKernelBase::emu_vgatherdps(const Xbyak::Xmm& xmm_val,
     const size_t kDataTypeSize = sizeof(float);
     Xbyak::Xmm xmm_mask{reg_mask.getIdx(), reg_mask.getKind(), static_cast<int>(reg_mask.getBit())};
     std::vector<Xbyak::Xmm> not_available_xmm{xmm_index, xmm_val, xmm_mask};
-    if (isValidIsa(x64::avx512_core)) {
+    if (is_valid_isa(x64::avx512_core)) {
         const Xbyak::Zmm zmm_zero_val = registersPool->getInplaceFree<Xbyak::Zmm>(not_available_xmm);
         push(zmm_zero_val);
         uni_vxorps(zmm_zero_val, zmm_zero_val, zmm_zero_val);
@@ -203,7 +44,7 @@ void JitKernelBase::emu_vgatherdps(const Xbyak::Xmm& xmm_val,
         vpcmpud(avx512_mask, Xbyak::Zmm{reg_mask.getIdx()}, zmm_zero_val, VCMPPS_GT);
         pop(zmm_zero_val);
         vgatherdps(xmm_val | avx512_mask, ptr[reg_addr + xmm_index * scale + disp]);
-    } else if (isValidIsa(x64::avx2)) {
+    } else if (is_valid_isa(x64::avx2)) {
         assert(reg_mask.isYMM());
         Xbyak::Ymm ymm_mask{reg_mask.getIdx()};
         vgatherdps(xmm_val, ptr[reg_addr + xmm_index * scale + disp], ymm_mask);
@@ -248,7 +89,7 @@ void JitKernelBase::emu_vgatherdps(const Xbyak::Xmm& xmm_val,
     }
 }
 
-void JitKernelBase::emu_vscatterdps(const Xbyak::Reg64& reg_addr,
+void jit_kernel_base::emu_vscatterdps(const Xbyak::Reg64& reg_addr,
                                     const Xbyak::Xmm& xmm_index,
                                     const int scale,
                                     const int disp,
@@ -258,7 +99,7 @@ void JitKernelBase::emu_vscatterdps(const Xbyak::Reg64& reg_addr,
     const size_t kDataTypeSize = sizeof(float);
     Xbyak::Xmm xmm_mask{reg_mask.getIdx(), reg_mask.getKind(), static_cast<int>(reg_mask.getBit())};
     std::vector<Xbyak::Xmm> not_available_xmm{xmm_index, xmm_val, xmm_mask};
-    if (isValidIsa(x64::avx512_core)) {
+    if (is_valid_isa(x64::avx512_core)) {
         const Xbyak::Zmm zmm_zero_val = registersPool->getInplaceFree<Xbyak::Zmm>(not_available_xmm);
         push(zmm_zero_val);
         uni_vxorps(zmm_zero_val, zmm_zero_val, zmm_zero_val);
@@ -284,7 +125,7 @@ void JitKernelBase::emu_vscatterdps(const Xbyak::Reg64& reg_addr,
         push(idx);
         push(mask);
         push(val);
-        if (isValidIsa(x64::avx2)) {
+        if (is_valid_isa(x64::avx2)) {
             push(Xbyak::Ymm{xmm_mask_temp.getIdx()});
             push(Xbyak::Ymm{xmm_index_temp.getIdx()});
             push(Xbyak::Ymm{xmm_val_temp.getIdx()});
@@ -314,7 +155,7 @@ void JitKernelBase::emu_vscatterdps(const Xbyak::Reg64& reg_addr,
             }
         };
 
-        if (isValidIsa(x64::avx2)) {
+        if (is_valid_isa(x64::avx2)) {
             for (int i = 0; i < static_cast<int>(kYmmSimdWidth / kXmmSimdWidth); i++) {
                 vextracti128(xmm_mask_temp, Xbyak::Ymm{xmm_mask.getIdx()}, i);
                 vextracti128(xmm_index_temp, Xbyak::Ymm{xmm_index.getIdx()}, i);
@@ -326,7 +167,7 @@ void JitKernelBase::emu_vscatterdps(const Xbyak::Reg64& reg_addr,
         }
         L(scatter_fast_end);
 
-        if (isValidIsa(x64::avx2)) {
+        if (is_valid_isa(x64::avx2)) {
             pop(Xbyak::Ymm{xmm_val_temp.getIdx()});
             pop(Xbyak::Ymm{xmm_index_temp.getIdx()});
             pop(Xbyak::Ymm{xmm_mask_temp.getIdx()});
@@ -337,7 +178,7 @@ void JitKernelBase::emu_vscatterdps(const Xbyak::Reg64& reg_addr,
     }
 }
 
-void JitKernelBase::gatherdd(const Xbyak::Xmm&    vDst,
+void jit_kernel_base::gatherdd(const Xbyak::Xmm&    vDst,
                              const Xbyak::Reg64&  rSrcPtr,
                              const Xbyak::Xmm&    vSrcShift,
                              const Xbyak::Opmask& kReadMask,
@@ -354,7 +195,7 @@ void JitKernelBase::gatherdd(const Xbyak::Xmm&    vDst,
     vpgatherdd(vDst | kReadMask, ptr[rSrcPtr + vSrcShift]);
 }
 
-void JitKernelBase::gatherdd(const Xbyak::Xmm&   vDst,
+void jit_kernel_base::gatherdd(const Xbyak::Xmm&   vDst,
                              const Xbyak::Reg64& rSrcPtr,
                              const Xbyak::Xmm&   vSrcShift,
                              const Xbyak::Xmm&   vReadMask,
@@ -366,7 +207,7 @@ void JitKernelBase::gatherdd(const Xbyak::Xmm&   vDst,
     if (zeroFill)
         pxor(vDst, vDst); // Don't use vpxor. It zeros the rest of the YMM register.
 
-    if (isValidIsa(x64::avx2)) {
+    if (is_valid_isa(x64::avx2)) {
         if (!useMask)
             uni_vpcmpeqd(vReadMask, vReadMask, vReadMask);
 
@@ -392,7 +233,7 @@ void JitKernelBase::gatherdd(const Xbyak::Xmm&   vDst,
     }
 }
 
-void JitKernelBase::gatherdd(const Xbyak::Ymm&   vDst,
+void jit_kernel_base::gatherdd(const Xbyak::Ymm&   vDst,
                              const Xbyak::Reg64& rSrcPtr,
                              const Xbyak::Ymm&   vSrcShift,
                              const Xbyak::Ymm&   vReadMask,
@@ -401,7 +242,7 @@ void JitKernelBase::gatherdd(const Xbyak::Ymm&   vDst,
     if (vDst.getIdx() == vSrcShift.getIdx() || vDst.getIdx() == vReadMask.getIdx() || vSrcShift.getIdx() == vReadMask.getIdx()) {
         IE_THROW() << "Any pair of the index, mask, or destination registers cannot be the same.";
     }
-    if (isValidIsa(x64::avx2)) {
+    if (is_valid_isa(x64::avx2)) {
         if (!useMask)
             uni_vpcmpeqd(vReadMask, vReadMask, vReadMask);
         if (zeroFill)
@@ -423,94 +264,7 @@ void JitKernelBase::gatherdd(const Xbyak::Ymm&   vDst,
     }
 }
 
-void JitKernelBase::uni_vpbroadcastd(const Xbyak::Xmm &x, const Xbyak::Operand &op) {
-    if (isValidIsa(x64::avx2)) {
-        vpbroadcastd(x, op);
-    } else if (isValidIsa(x64::avx)) {
-        if (op.isMEM()) {
-            vbroadcastss(x, op.getAddress());
-        } else {
-            vmovss(x, x, op);
-            vpshufd(x, x, 0x0);
-        }
-    } else {
-        movss(x, op);
-        pshufd(x, x, 0x0);
-    }
-}
-
-void JitKernelBase::uni_vpbroadcastd(const Xbyak::Ymm &x, const Xbyak::Operand &op) {
-    if (isValidIsa(x64::avx2)) {
-        vpbroadcastd(x, op);
-    } else {
-        if (op.isMEM()) {
-            vbroadcastss(x, op.getAddress());
-        } else {
-            const Xbyak::Xmm t(x.getIdx());
-            if (!t.isEqualIfNotInherited(op)) {
-                vmovss(t, t, op);
-            }
-            vinsertf128(x, x, t, 1);
-            vshufps(x, x, x, 0);
-        }
-    }
-}
-
-void JitKernelBase::uni_vaddps(const Xbyak::Xmm& x, const Xbyak::Xmm& op1, const Xbyak::Operand& op2) {
-    if (isValidIsa(x64::avx)) {
-        vaddps(x, op1, op2);
-    } else {
-        if (x.getIdx() == op1.getIdx()) {
-            addps(x, op2);
-        } else if (x.isEqualIfNotInherited(op2)) {
-            addps(x, op1);
-        } else {
-            movups(x, op1);
-            addps(x, op2);
-        }
-    }
-}
-
-void JitKernelBase::uni_vsubps(const Xbyak::Xmm& x, const Xbyak::Xmm& op1, const Xbyak::Operand& op2) {
-    if (isValidIsa(x64::avx)) {
-        vsubps(x, op1, op2);
-    } else {
-        if (x.getIdx() == op1.getIdx()) {
-            subps(x, op2);
-        } else if (x.isEqualIfNotInherited(op2)) {
-            push(op1);
-            subps(op1, op2);
-            movups(x, op1);
-            pop(op1);
-        } else {
-            movups(x, op1);
-            subps(x, op2);
-        }
-    }
-}
-
-void JitKernelBase::uni_vcmpps(const Xbyak::Xmm& x,
-                               const Xbyak::Xmm& op1,
-                               const Xbyak::Operand& op2,
-                               const int cmp_predicate) {
-    if (isValidIsa(x64::avx)) {
-        vcmpps(x, op1, op2, cmp_predicate);
-    } else {
-        if (x.getIdx() == op1.getIdx()) {
-            cmpps(x, op2, cmp_predicate);
-        } else if (x.isEqualIfNotInherited(op2)) {
-            push(op1);
-            cmpps(op1, op2, cmp_predicate);
-            movups(x, op1);
-            pop(op1);
-        } else {
-            movups(x, op1);
-            cmpps(x, op2, cmp_predicate);
-        }
-    }
-}
-
-void JitKernelBase::fillRestWorkMask(const Xbyak::Opmask& dstMask,
+void jit_kernel_base::fillRestWorkMask(const Xbyak::Opmask& dstMask,
                                      const Xbyak::Zmm&    zAux,
                                      const Xbyak::Reg64&  rWorkRest) {
     auto rAux0 = getReg64();
@@ -533,171 +287,7 @@ void JitKernelBase::fillRestWorkMask(const Xbyak::Opmask& dstMask,
     kmovw(dstMask, rOnes);
 }
 
-void JitKernelBase::load(const Xbyak::Xmm&     vDst,
-                         const Xbyak::Address& srcAddr,
-                         const Xbyak::Reg64&   rLoadNum,
-                         const size_t          typeSize,
-                         const bool            zeroFilling) {
-    if (!one_of(typeSize, 1, 2, 4, 8)) {
-        IE_THROW() << "Could not load data with type size " << typeSize;
-    }
-    const uint8_t elPerVec = x64::cpu_isa_traits<x64::sse41>::vlen / typeSize;
-    Xbyak::Label lEnd;
-    if (zeroFilling)
-        pxor(vDst, vDst);
-
-    for (uint8_t i = 0; i < elPerVec; i++) {
-        cmp(rLoadNum, i);
-        jle(lEnd, T_NEAR);
-
-        const size_t offset = i * typeSize;
-        if (typeSize == 1)
-            pinsrb(vDst, ptr[srcAddr.getRegExp() + offset], i);
-        else if (typeSize == 2)
-            pinsrw(vDst, ptr[srcAddr.getRegExp() + offset], i);
-        else if (typeSize == 4)
-            pinsrd(vDst, ptr[srcAddr.getRegExp() + offset], i);
-        else if (typeSize == 8)
-            pinsrq(vDst, ptr[srcAddr.getRegExp() + offset], i);
-    }
-    L(lEnd);
-}
-
-void JitKernelBase::load(const Xbyak::Ymm&     vDst,
-                         const Xbyak::Address& srcAddr,
-                         const Xbyak::Reg64&   rLoadNum,
-                         const size_t          typeSize,
-                         const bool            zeroFilling) {
-    if (!one_of(typeSize, 1, 2, 4, 8)) {
-        IE_THROW() << "Could not load data with type size " << typeSize;
-    }
-    const size_t elPerXmm = x64::cpu_isa_traits<x64::sse41>::vlen / typeSize;
-    Xbyak::Label lEnd;
-    if (zeroFilling)
-        uni_vpxor(vDst, vDst, vDst);
-    Xbyak::Xmm xmmDst(vDst.getIdx());
-
-    for (size_t i = 0lu; i < 2lu; i++) {
-        Xbyak::Label lPerm;
-        const size_t idx = i * elPerXmm;
-        const size_t offset0 = idx * typeSize;
-
-        for (size_t j = 0lu; j < elPerXmm; j++) {
-            cmp(rLoadNum, j + idx);
-            jle(i == 0 ? lEnd : lPerm, T_NEAR);
-
-            const size_t offset = offset0 + j * typeSize;
-            if (typeSize == 1)
-                pinsrb(xmmDst, ptr[srcAddr.getRegExp() + offset], j);
-            else if (typeSize == 2)
-                pinsrw(xmmDst, ptr[srcAddr.getRegExp() + offset], j);
-            else if (typeSize == 4)
-                pinsrd(xmmDst, ptr[srcAddr.getRegExp() + offset], j);
-            else if (typeSize == 8)
-                pinsrq(xmmDst, ptr[srcAddr.getRegExp() + offset], j);
-        }
-
-        L(lPerm);
-        vperm2f128(vDst, vDst, vDst, 0x1);
-    }
-    L(lEnd);
-}
-
-void JitKernelBase::store(const Xbyak::Address& dstAddr,
-                          const Xbyak::Xmm&     vSrc,
-                          const Xbyak::Reg64&   rToStoreNum,
-                          const size_t          typeSize) {
-    if (!one_of(typeSize, 1, 2, 4, 8)) {
-        IE_THROW() << "Could not store data with type size " << typeSize;
-    }
-    Xbyak::Label lEnd;
-    const size_t elPerVec = x64::cpu_isa_traits<x64::sse41>::vlen / typeSize;
-
-    for (size_t i = 0; i < elPerVec; i++) {
-        cmp(rToStoreNum, i);
-        jle(lEnd, T_NEAR);
-
-        const size_t offset = i * typeSize;
-        if (typeSize == 1) {
-            uni_vpextrb(ptr[dstAddr.getRegExp() + offset], vSrc, i);
-        } else if (typeSize == 2) {
-            uni_vpextrw(ptr[dstAddr.getRegExp() + offset], vSrc, i);
-        } else if (typeSize == 4) {
-            uni_vpextrd(ptr[dstAddr.getRegExp() + offset], vSrc, i);
-        } else if (typeSize == 8) {
-            uni_vpextrq(ptr[dstAddr.getRegExp() + offset], vSrc, i);
-        }
-    }
-    L(lEnd);
-}
-
-void JitKernelBase::store(const Xbyak::Address& dstAddr,
-                          const Xbyak::Ymm&     vSrc,
-                          const Xbyak::Reg64&   rToStoreNum,
-                          const size_t          typeSize) {
-    if (!one_of(typeSize, 1, 2, 4, 8)) {
-        IE_THROW() << "Could not store data with type size " << typeSize;
-    }
-    Xbyak::Label lEnd;
-    Xbyak::Xmm xmmSrc(vSrc.getIdx());
-    const size_t elPerXmm = x64::cpu_isa_traits<x64::sse41>::vlen / typeSize;
-
-    for (int i = 0; i < 2; i++) {
-        Xbyak::Label lPerm;
-        const size_t idx = i * elPerXmm;
-        const size_t offset0 = idx * typeSize;
-
-        for (size_t j = 0; j < elPerXmm; j++) {
-            cmp(rToStoreNum, j + idx);
-            jle(i == 0 ? lEnd : lPerm, T_NEAR);
-
-            const size_t offset = offset0 + j * typeSize;
-            if (typeSize == 8) {
-                uni_vpextrq(ptr[dstAddr.getRegExp() + offset], xmmSrc, j);
-            } else if (typeSize == 4) {
-                uni_vpextrd(ptr[dstAddr.getRegExp() + offset], xmmSrc, j);
-            } else if (typeSize == 2) {
-                uni_vpextrw(ptr[dstAddr.getRegExp() + offset], xmmSrc, j);
-            } else if (typeSize == 1) {
-                uni_vpextrb(ptr[dstAddr.getRegExp() + offset], xmmSrc, j);
-            }
-        }
-
-        L(lPerm);
-        vperm2f128(vSrc, vSrc, vSrc, 0x1);
-    }
-    L(lEnd);
-}
-
-
-void JitKernelBase::push(const Xbyak::Xmm& xmm) {
-    if (xmm.isXMM()) {
-        sub(rsp, xmm_len);
-        uni_vmovdqu(ptr[rsp], xmm);
-    } else if (xmm.isYMM()) {
-        sub(rsp, ymm_len);
-        uni_vmovdqu(ptr[rsp], Xbyak::Ymm{xmm.getIdx()});
-    } else if (xmm.isZMM()) {
-        sub(rsp, zmm_len);
-        uni_vmovdqu(ptr[rsp], Xbyak::Zmm{xmm.getIdx()});
-    }
-}
-
-void JitKernelBase::pop(const Xbyak::Xmm& xmm) {
-    if (xmm.isXMM()) {
-        uni_vmovdqu(xmm, ptr[rsp]);
-        add(rsp, xmm_len);
-    } else if (xmm.isYMM()) {
-        uni_vmovdqu(Xbyak::Ymm{xmm.getIdx()}, ptr[rsp]);
-        add(rsp, ymm_len);
-    } else if (xmm.isZMM()) {
-        uni_vmovdqu(Xbyak::Zmm{xmm.getIdx()}, ptr[rsp]);
-        add(rsp, zmm_len);
-    }
-}
-
-
-void JitKernelBase::memMovDD(const Xbyak::Reg64& rDst,
+void jit_kernel_base::memMovDD(const Xbyak::Reg64& rDst,
                              const Xbyak::Reg64& rSrc,
                              const Xbyak::Xmm&   vReadMask,
                              const Xbyak::Xmm&   vSrcShift,
@@ -737,7 +327,7 @@ void JitKernelBase::memMovDD(const Xbyak::Reg64& rDst,
     L(lEnd);
 }
 
-void JitKernelBase::memMovDD(const Xbyak::Reg64& rDst,
+void jit_kernel_base::memMovDD(const Xbyak::Reg64& rDst,
                              const Xbyak::Reg64& rSrc,
                              const Xbyak::Ymm&   vReadMask,
                              const Xbyak::Ymm&   vSrcShift,
@@ -745,11 +335,11 @@ void JitKernelBase::memMovDD(const Xbyak::Reg64& rDst,
                              const bool          useMask,
                              const bool          zeroFill) {
     Xbyak::Label lEnd;
-    if (isValidIsa(x64::avx2)) {
+    if (is_valid_isa(x64::avx2)) {
         auto vAux = RegistersPool::Reg<Xbyak::Ymm>(registersPool);
         gatherdd(vAux, rSrc, vSrcShift, vReadMask, useMask, zeroFill);
         store(ptr[rDst], vAux, rToStoreNum, sizeof(int));
-    } else if (isValidIsa(x64::avx)) {
+    } else if (is_valid_isa(x64::avx)) {
         const uint8_t typeSize = sizeof(int);
         const uint8_t elPerXmm = x64::cpu_isa_traits<x64::sse41>::vlen / typeSize;
         Xbyak::Xmm xmmReadMask  = Xbyak::Xmm(vReadMask.getIdx()),
@@ -775,10 +365,10 @@ void JitKernelBase::memMovDD(const Xbyak::Reg64& rDst,
     L(lEnd);
 }
 
-void JitKernelBase::createRegistersPool() {
-    registersPool = RegistersPool::create(maxCpuIsa, {abi_param1});
+void jit_kernel_base::createRegistersPool() {
+    registersPool = RegistersPool::create(max_cpu_isa_, {abi_param1});
 }
 
-void JitKernelBase::createStackAllocator() {
+void jit_kernel_base::createStackAllocator() {
     stackAllocator = std::unique_ptr<StackAllocator>(new StackAllocator{*this});
 }
